Use relative tab domain wrapper exports
